Updated priority of destructors in first round. Also including compiled versions 1-4.
V1 is a baseline integration of the switching mechanism.
V2 is a completed integration of both defensive and attacking switching.
V3 is a version of V2 with a updated priority for building defenses before tunnel.
V4 is a version of V2/V3 with updated priorty for destructor placement at start of game.